[Feat] Reservation: serviceImpl 구현 #23
